requirements.txt
Flask==3.03
requests==2.32.3
python-dotenv==1.01